requirements.txt
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab_server
matplotlib
matplotlib-inline
notebook
notebook_shim
numpy
pandas
plotly
scikit-learn
scipy
seaborn
sklearn-compat
statsmodels
streamlit
streamlit-javascript
xgboost
